+++ src/compas_viewer/actions/action.py
@@ -57,7 +57,7 @@ class Action(QObject):
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
