+++ src/compas_viewer/components/renderer/camera.py
@@ -163,7 +163,7 @@ class Camera:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
