+++ src/compas_viewer/components/renderer/selector.py
@@ -65,7 +65,7 @@ class Selector(QObject):
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
