+++ src/compas_viewer/components/treeform_components.py
@@ -18,7 +18,7 @@ class TreeformComponents:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
