+++ src/compas_viewer/controller/controller.py
@@ -46,7 +46,7 @@ class Controller:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
