+++ src/compas_viewer/scene/geometryobject.py
@@ -73,7 +73,7 @@ class GeometryObject(ViewerSceneObject, BaseGeometryObject):
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
