+++ src/compas_viewer/scene/meshobject.py
@@ -94,7 +94,7 @@ class MeshObject(ViewerSceneObject, BaseMeshObject):
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
