+++ src/compas_viewer/ui/mainwindow.py
@@ -6,7 +6,7 @@ from PySide6.QtWidgets import QWidget
 class MainWindow(QMainWindow):
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
