+++ src/compas_viewer/ui/menubar.py
@@ -14,7 +14,7 @@ class MenuBar:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
