+++ src/compas_viewer/ui/statusbar.py
@@ -8,7 +8,7 @@ class SatusBar:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
