+++ src/compas_viewer/ui/toolbar.py
@@ -11,7 +11,7 @@ class ToolBar:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
