+++ src/compas_viewer/ui/ui.py
@@ -15,7 +15,7 @@ class UI:
 
     @property
     def viewer(self):
-        from compas_viewer.main import Viewer
+        from compas_viewer.viewer import Viewer
 
         return Viewer()
